Update CEL, fix tests, clean up code
