+++ app/src/main/java/com/superwall/superapp/MainApplication.kt
@@ -76,7 +76,7 @@ class MainApplication :
                 SuperwallOptions().apply {
                     paywalls =
                         PaywallOptions().apply {
-                            shouldPreload = true
+                            shouldPreload = false
                         }
                 },
         )
