+++ build.gradle.kts
@@ -5,7 +5,6 @@ plugins {
     alias(libs.plugins.androidApplication) apply false
     alias(libs.plugins.kotlinAndroid) apply false
     alias(libs.plugins.serialization) apply false
-    alias(libs.plugins.jetbrainsKotlinJvm) apply false
 }
 true
