+++ settings.gradle
@@ -18,4 +18,3 @@ include ':app'
 include ':superwall'
 include 'example:app'
 project(':example:app').projectDir = new File('example/app')
-include ':supercel'
